+++ src/components/app-header/app-header.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
 import { AppHeaderUI } from '../ui';
-import React from 'react';
+import { links } from '@utils-constants';
 
-export const AppHeader: FC = () => <AppHeaderUI userName='' />;
+export const AppHeader: FC = () => <AppHeaderUI links={links} />;