+++ src/components/app/app.tsx
@@ -1,7 +1,26 @@
-import clsx from 'clsx';
-import { AppHeader } from '../app-header';
-import { Modal } from '../modal';
-import { FormInput } from '../ui/form-input-prompt';
-import React from 'react';
+import { AppHeader } from '@components';
+import style from './app.module.scss';
+import { AboutMePage } from '@pages';
+import { Route, Routes, useLocation } from 'react-router-dom';
+import { NotFound404 } from '../../pages';
 
-export const App = () => <main />;
+// TODO: изменить ссылки
+
+export const App = () => {
+	const location = useLocation();
+
+	return (
+		<div className={style.app}>
+			<AppHeader />
+			<Routes location={location}>
+				<Route path='/' element={<AboutMePage />} />
+
+				{/* <Route path='/' element={<SectionAboutMe />} />
+				<Route path='/stack' element={<SectionStack />} />
+				<Route path='/portfolio' element={<SectionPortfolio />} /> */}
+
+				<Route path='*' element={<NotFound404 />} />
+			</Routes>
+		</div>
+	);
+};