+++ src/components/common/icon/icon.tsx
@@ -1,0 +1,5 @@
+import { FC } from 'react';
+import { IconUI } from '@ui/index';
+import { TIconProps } from './type';
+
+export const Icon: FC<TIconProps> = ({ links }) => <IconUI links={links} />;