+++ src/components/common/icon/type.ts
@@ -1,0 +1,9 @@
+export type TIconLinks = {
+	email?: string;
+	github?: string;
+	telegram?: string;
+};
+
+export type TIconProps = {
+	links: TIconLinks;
+};