+++ src/components/common/link/type.ts
@@ -1,0 +1,4 @@
+export type TLinkProps = {
+	link: string;
+	text: string;
+};