+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './app-header';
+export * from './modal';
+export * from './common';