+++ src/components/ui/cart-quadcopter/type.ts
@@ -1,9 +1,0 @@
-type CardQuadcopter = {
-	title: string;
-	bodyText: string;
-	buttonText: string;
-	anchorLink: string;
-	modalData: any;
-};
-
-export default CardQuadcopter;