+++ src/components/ui/common/icon/type.ts
@@ -1,0 +1,9 @@
+export type TIconDataLinks = {
+	email?: string;
+	github?: string;
+	telegram?: string;
+};
+
+export type TIconUIProps = {
+	links: TIconDataLinks;
+};