+++ src/components/ui/common/link/type.ts
@@ -1,0 +1,4 @@
+export type TLinkUIProps = {
+	link: string;
+	text: string;
+};