+++ src/components/ui/index.ts
@@ -3,4 +3,4 @@ export * from './modal-overlay';
 export * from './pages';
 export * from './app-header';
 export * from './modal';
-export * from './profile-menu';
+export * from './common';