+++ src/components/ui/pages/index.ts
@@ -1,4 +1,2 @@
-export * from './forms-page';
-export * from './login';
-export * from './profile';
-export * from './register';
+export * from './about-me';
+export * from './common';