+++ src/pages/about-me/about-me.tsx
@@ -1,0 +1,3 @@
+import { AboutMePageUI } from '@ui/index';
+
+export const AboutMePage = () => <AboutMePageUI />;