+++ src/pages/common/type.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+
+export type TCommonPage = {
+	pageStyles: string;
+	children: ReactNode;
+};