+++ src/pages/index.ts
@@ -1,1 +1,3 @@
 export * from './not-fount-404';
+export * from './about-me';
+export * from './common';