+++ src/utils/constants.ts
@@ -1,0 +1,5 @@
+export const links = {
+	email: 'dorohev.kir@gmail.com',
+	github: 'https://github.com/DKMFZF',
+	telegram: 'https://t.me/DKMFZF'
+};