+++ webpack/webpack.common.js
@@ -65,7 +65,12 @@ module.exports = {
 					{
 						loader: 'sass-loader',
 						options: {
-							sourceMap: true
+							sourceMap: true,
+							sassOptions: {
+								includePaths: [
+									path.resolve(__dirname, '..', 'src/styles')
+								]
+							}
 						}
 					}
 				]
@@ -75,9 +80,37 @@ module.exports = {
 	resolve: {
 		extensions: ['.js', '.jsx', '.tsx', '.ts', '.json'],
 		alias: {
-			fonts: path.resolve(__dirname, '..', './src/fonts'),
 			src: path.resolve(__dirname, '..', './src'),
-			components: path.resolve(__dirname, '..', './src/components')
+			styles: path.resolve(__dirname, '..', 'src/styles'),
+			// path.resolve(__dirname, '..', 'src/styles/'),
+			fonts: path.resolve(__dirname, '..', './src/fonts'),
+			components: path.resolve(__dirname, '..', './src/components'),
+			'@pages': path.resolve(__dirname, '..', './src/pages'),
+			'@components': path.resolve(__dirname, '..', './src/components'),
+			'@ui': path.resolve(__dirname, '..', './src/components/ui'),
+			'@ui-pages': path.resolve(
+				__dirname,
+				'..',
+				'./src/components/ui/pages'
+			),
+			'@utils-types': path.resolve(__dirname, '..', './src/utils/types'),
+			'@utils-constants': path.resolve(
+				__dirname,
+				'..',
+				'./src/utils/constants'
+			),
+			'@api': path.resolve(
+				__dirname,
+				'..',
+				'./src/utils/flv-test-api.ts'
+			),
+			'@slices': path.resolve(__dirname, '..', './src/services/slices'),
+			'@selectors': path.resolve(
+				__dirname,
+				'..',
+				'./src/services/selectors'
+			),
+			'@images': path.resolve(__dirname, '..', './src/image')
 		}
 	},
 	plugins: [
@@ -93,5 +126,13 @@ module.exports = {
 		new webpack.EnvironmentPlugin({
 			NODE_ENV: 'development'
 		})
-	]
+	],
+	optimization: production
+		? {
+				splitChunks: {
+					chunks: 'all'
+				},
+				runtimeChunk: 'single'
+		  }
+		: undefined
 };