+++ CHANGELOG.md
@@ -2,6 +2,105 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.26.1](https://github.com/DKMFzF/frontend-portfolio/compare/v1.26.0...v1.26.1) (2025-04-20)
+
+## [1.26.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.25.0...v1.26.0) (2025-04-20)
+
+### Features
+
+-   **section-stack:** add new page ([0faf69a](https://github.com/DKMFzF/frontend-portfolio/commit/0faf69a7c02fb064d032f25118cb9e64628f87da))
+
+## [1.25.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.24.0...v1.25.0) (2025-04-20)
+
+### Features
+
+-   **section=portfolio:** add new page ([63ac502](https://github.com/DKMFzF/frontend-portfolio/commit/63ac50247313b3db7c1fcdd754ceebee643f2530))
+
+## [1.24.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.23.0...v1.24.0) (2025-04-20)
+
+### Features
+
+-   **app:** update route ([3c6ddd7](https://github.com/DKMFzF/frontend-portfolio/commit/3c6ddd787ed6cefea8374ca62da1f3a5867a9e68))
+
+## [1.23.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.22.0...v1.23.0) (2025-04-20)
+
+### Features
+
+-   **section-about-me:** add main version ([cd240c1](https://github.com/DKMFzF/frontend-portfolio/commit/cd240c15660a275a7dac5789342fd5b0ab298686))
+
+## [1.22.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.21.0...v1.22.0) (2025-04-19)
+
+### Features
+
+-   **section-about-me:** add first version section ([c7a8692](https://github.com/DKMFzF/frontend-portfolio/commit/c7a86929b3a211a91821cbc37453f73f092fd6d4))
+
+## [1.21.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.20.0...v1.21.0) (2025-04-19)
+
+### Features
+
+-   **section-container:** add wrapper section ([c350c61](https://github.com/DKMFzF/frontend-portfolio/commit/c350c61a46113aadd7acd07fc7cecdfd80bfbc52))
+
+## [1.20.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.19.0...v1.20.0) (2025-04-19)
+
+### Features
+
+-   **icon:** add wrapper ([4cdddc2](https://github.com/DKMFzF/frontend-portfolio/commit/4cdddc200f9d1ca4cc9d4eea9fbe2c956d15768b))
+
+## [1.19.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.17.0...v1.19.0) (2025-04-19)
+
+### Features
+
+-   **header:** update link-icon ([52b0705](https://github.com/DKMFzF/frontend-portfolio/commit/52b07058e36236a6ae869574f4e7e03357fdd513))
+-   **test:** asd ([e2fcc12](https://github.com/DKMFzF/frontend-portfolio/commit/e2fcc12eb40ae2559ec500598a554bf42fa214cc))
+
+## [1.18.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.17.0...v1.18.0) (2025-04-19)
+
+### Features
+
+-   **header:** update link-icon ([52b0705](https://github.com/DKMFzF/frontend-portfolio/commit/52b07058e36236a6ae869574f4e7e03357fdd513))
+
+## [1.17.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.16.0...v1.17.0) (2025-04-19)
+
+### Features
+
+-   **link:** add component link ([75aa14f](https://github.com/DKMFzF/frontend-portfolio/commit/75aa14f6e4d047cbe5a825590a39c6868e9aa199))
+
+## [1.16.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.15.0...v1.16.0) (2025-04-19)
+
+### Features
+
+-   **icons:** refactor svg-code add min size ([5efd51b](https://github.com/DKMFzF/frontend-portfolio/commit/5efd51b57e3280d826c0d056b4838818ca463fbc))
+
+## [1.15.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.14.0...v1.15.0) (2025-04-19)
+
+### Features
+
+-   **font:** add new font Iter ([a355d17](https://github.com/DKMFzF/frontend-portfolio/commit/a355d176ccc66862a297e2953b829c5cfdba0932))
+
+## [1.14.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.13.0...v1.14.0) (2025-04-19)
+
+### Features
+
+-   **font:** add new font Playfair ([c236e11](https://github.com/DKMFzF/frontend-portfolio/commit/c236e11bb5cd2ad88f26d3368c48159c7d3b785c))
+
+## [1.13.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.12.0...v1.13.0) (2025-04-19)
+
+### Features
+
+-   **webpack:** add import var setting in config ([1019927](https://github.com/DKMFzF/frontend-portfolio/commit/10199276daaeedb90fa57a18f8f86d596ce142b4))
+
+## [1.12.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.11.0...v1.12.0) (2025-04-19)
+
+### Features
+
+-   **icon:** add component icon ([10f043d](https://github.com/DKMFzF/frontend-portfolio/commit/10f043d431977a4b19042c2558560de57b8844ca))
+
+## [1.11.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.10.0...v1.11.0) (2025-04-19)
+
+### Features
+
+-   **tscomfig:** add file in project ([da86684](https://github.com/DKMFzF/frontend-portfolio/commit/da866840f6a458defa878aec8a6c9b901378fca5))
+
 ## [1.10.0](https://github.com/DKMFzF/frontend-portfolio/compare/v1.9.0...v1.10.0) (2025-04-19)
 
 ### Features